backend/requirements.txt
absl-py==1.1.0
aiohttp==3.8.3
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
datasets==2.8.0
dill==0.3.6
distlib==0.3.6
dnspython==2.3.0
email-validator==2.0.0.post2
filelock==3.9.0
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-WTF==1.1.1
flatbuffers==1.12
frozenlist==1.3.3
fsspec==2022.11.0
gast==0.4.0
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.7.0
huggingface-hub==0.11.1
idna==3.3
itsdangerous==2.1.2
javalang==0.13.0
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
mongoengine==0.27.0
MouseInfo==0.1.3
multidict==6.0.4
multiprocess==0.70.14
numpy
oauthlib==3.2.0
openai==0.26.5
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
Pillow==9.1.1
platformdirs==3.1.0
protobuf==3.19.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.7.0
pymongo==4.3.3

